Update workflow configuration in dotnet.yml
The changes in this commit update the 'dotnet.yml' github workflow file. Specifically, we've introduced a 'fetch-depth' of 0 in the checkout step for a complete history for all branches and tags. Additionally, the 'dotnet-releaser' command in the 'build, tests, cover, pack, and publish' step has been changed from 'publish' to 'run'.